requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
backcall==0.2.0
certifi==2021.5.30
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.1.0
heapq-max==0.21
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.5
openpyxl==3.0.7
pandas==1.2.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2021.5.29
prompt-toolkit==3.0.10
psycopg2-binary==2.9.1
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pyzmq==20.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
